#2156 mv substitution functionality into separate class
